+++ docs/conf.py
@@ -49,7 +49,7 @@ project = 'IDPConformerGenerator'
 year = '2022'
 author = 'Julie Forman-Kay Lab'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '0.7.7'
+version = release = '0.7.8'
 
 pygments_style = 'trac'
 templates_path = ['.']