+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ While in version ``0``, minor and patch upgrades converge in the ``patch`` numbe
 Changelog
 =========
 
+v0.7.8 (2023-10-02)
+------------------------------------------------------------
+
 * Bug-fix in ``ldrs`` for input sequence recognition
 
 v0.7.7 (2023-09-26)